refactor lighting into its own files
